Implemented authentication (#2)
* Added new user controller
Along with its entity and an internal entity that delivers the status of the process

* Renamed interface to better reflect its usage

* Added controller to check if a user exists
It also checks if the user has their email verified. In case this is false, the user is considered as non-existing

* Created controller to check user credentials and deliver token to frontend

* Added extra information to login controller
Now we know the difference between a bad username/password combination and a server error

* Added email verification system
At the moment we send an ugly, unformatted message. We plan to replace it with a HTML message with the project's logo

* Added controller to send email with password restoration link
We are missing the actual restoration controller. I'll later add a function to create links with verification data for both the password restoration and the email verification controllers

* Lowered db connections to avoid overloading
We're using a free db that only allows 10 connections at a time, I don't want Heroku taking half of that

* Added packages for request validation

* Created auth router and added basic registration route
Next commits I will be adding validation and authentication to routers

* Declared validation conditions on classes

* Added validation to registration route
Next commit this implementation will be converted to a middleware

* Moved validation logic to a middleware
Now we can validate any class in any route, you just need to provide the type and we return the parsed data in the res.locals object

* Added email template for "password forgotten" messages

* Added configuration to use email template and embed logo

* Added validation entity for the forgot password route

* Added 404 response when email does not exist on the database
Wondering if this is a security vulnerability, but for now this is staying

* Added URL constructor using JWT

* Added verify email template

* Added login route

* Added support for URL parameters on the validator middleware

* Changed email verification controller to work with JWT
We're missing the implementation to invalidate the database's token at this point, so this route is not usable yet

* Added token invalidator

* Documented registration route

* Documented password forgotten route
I'll later add the missing entities

* Documented login route

* Documented email verification route

* Added missing entities' documentation

* Added validation data entity with its docs